requirements.txt
mkdocs==1.6.*
mkdocs-material[imaging]==9.*
mkdocs-awesome-nav==3.*
mkdocs-macros-plugin==1.3.*
PyYAML==6.*